chore: split large code file in to maintainable files
